feat: Refactor ModelServingRepository and ModelServingService for improved endpoint handling and session management
